Add toggler for switch data sources
