+++ .env.default
@@ -1,17 +1,17 @@
 NEXT_PUBLIC_ADRESSE_URL=http://localhost:3000
 
-# --------------------------------------------------------------------------------------------------------------
-#  Parametres URL Externe autre projet BAN
-# ---------------------------------------------------------------------------------------------------------------
+# -----------------------------------------------
+# --- Parametres URL Externe autre projet BAN ---
+# -----------------------------------------------
 NEXT_PUBLIC_API_GEO_URL=https://geo.api.gouv.fr/2023
 NEXT_PUBLIC_API_ETABLISSEMENTS_PUBLIC=https://api-lannuaire.service-public.fr/api/explore/v2.1
 
 NEXT_PUBLIC_API_BAN_URL=https://plateforme.adresse.data.gouv.fr
 NEXT_PUBLIC_API_ADRESSE=https://api-adresse.data.gouv.fr
 
-# --------------------------------------------------------------------------------------------------------------
-#  Utilisation appel externe Bal
-# ---------------------------------------------------------------------------------------------------------------
+# -------------------------------------
+# --- Utilisation appel externe Bal ---
+# -------------------------------------
 
 NEXT_PUBLIC_API_DEPOT_URL=https://plateforme-bal.adresse.data.gouv.fr/api-depot
 API_DEPOT_TOKEN=
@@ -28,29 +28,32 @@ NEXT_PUBLIC_BAL_ADMIN_API_URL=https://bal-admin.adresse.data.gouv.fr/api
 NEXT_PUBLIC_BAL_API_URL=https://api-bal.adresse.data.gouv.fr/v1
 NEXT_PUBLIC_MES_ADRESSES=https://mes-adresses.data.gouv.fr
 
-# --------------------------------------------------------------------------------------------------------------
-#  Utilisation Matomo pour les stats
-# ---------------------------------------------------------------------------------------------------------------
+# -----------------------------------------
+# --- Utilisation Matomo pour les stats ---
+# -----------------------------------------
 NEXT_PUBLIC_MATOMO_URL=
 NEXT_PUBLIC_MATOMO_SITE_ID=
-MATOMO_TOKEN_AUTH=
+# The Matomo Token is actualy not required
+# MATOMO_TOKEN_AUTH=
 
 # A voir pour supprimer cette variable qui utilise la lib helmet
 ENABLE_HELMET=0
 
-# Defintion du port 
+# Defintion du port
 PORT=3000
 
 # Valeur 0 ou 1 pour la demande de certification
 NEXT_PUBLIC_CERTIFICAT_NUMEROTATION_ENABLED=
 
+# --------------------------------------
+# --- Connexion au S3 pour les datas ---
+# --------------------------------------
+
+S3_CONFIG_ACCESS_KEY_ID=
+S3_CONFIG_SECRET_ACCESS_KEY=
+S3_CONFIG_REGION=
+S3_CONFIG_ENDPOINT=
+
+# Connexion au Dossier Data local (en cas de  rollback - TOGGLER_DATA_SOURCES=local sinon TOGGLER_DATA_SOURCES=S3)
+TOGGLER_DATA_SOURCES=S3
 PATH_STATIC_FILE=
-
-# --------------------------------------------------------------------------------------------------------------
-#  Connexion au S3 pour les datas
-# ---------------------------------------------------------------------------------------------------------------
-
-S3_CONFIG_ACCESS_KEY_ID =
-S3_CONFIG_SECRET_ACCESS_KEY =
-S3_CONFIG_REGION =
-S3_CONFIG_ENDPOINT =
